power: qpnp-vm-bms: Fix the catch-up time calculation
The catch-up time is calculated when the device moves
from not-charging to charging state. There is possibility
when the new catch-up SOC may get updated after the
catch-up time is calculated, leading to a quicker
SOC ramp-up (than expected). Fix this by resetting
the catch-up time when SOC changes and recalculating
it if the SOC gap is >= MIN_SOC_UUC.

CRs-Fixed: 788553
Change-Id: Id35c6a160eb385dea8792c062c32ab0571d467d7
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>