power: qpnp-vm-bms: Apply UUC logic only while discharging
UUC (unusable capacity) is applied when a charger is present
and battery is discharging. Change it to be applied only when
we are discharging.

In some cases, even with the charger present the discharge
current > charge current (battery is net discharging). In
this case, the SOC may not reach 0 at the cutoff voltage
(as the UUC reduction is not applied). This can cause the
battery voltage to go critically low causing a UVLO.

CRs-Fixed: 793706
Change-Id: Ibe21bdfeb82565816e4555d60b627c9f2adff099
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>