Revert "usb: dwc3: otg: Disable host mode runtime PM"
This reverts commit a306ab2b8e5651619e6054e8398a3bfc0795f834.
The root cause of the original problem of the PHY not entering
L2 was identified as coming from the xHCI stack writing the
PED bit of the PORTSC register when the port was already
disabled. This causes the controller to bring the PHY out of
L2 and not suspend it again. This is fixed in commit e0f2206cd8
"usb: host: xhci: Do not disable already disabled HS root
hub port". With that fix in place, we no longer need to work
around the issue and can re-enable runtime PM.

Change-Id: I5a8b97f396a706f739275f17b84924f40a30df71
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: GreyLeshy <andreiskull@gmail.com>