soc: qcom: glink_smd_xprt: Remove erroneous uses of mutex
The mutex protecting the in_ssr flag in the edge information structure is
locked and unlocked in places where this flag is already protected by an
RCU, and in functions which may be called in atomic context.

Remove these uses of the mutex. Protection of the flag is maintained by the
RCU in these cases.

Change-Id: I2be88be0b42fd403437ff53d9d6acd7f1c3f4fa1
Signed-off-by: Steven Cahail <scahail@codeaurora.org>