mtd: msm_qpic_nand: Return EUCLEAN only when max bits are corrected
In the nand devices the pages are now allowed to have bitflips even
on the first ever write to the page. Thus reporting EUCLEAN on an
arbitrary number like 3 bitflips per codeword would result in a
large number of EUCLEAN errors reported.

Change the behaviour of the driver to return EUCLEAN only when the
ECC bits corrected in a codeword equal the max number that can
be corrected.

Change-Id: I914aa93f0752487a7b79a7cb6afecb5ce9f9ab61
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>