media: dvb: delay processing thread during control operations
Processing thread might take a relatively long time to process
a buffer for SDMX. During this time, control operations such as
adding or removing a channel must wait for the mutex to be unlocked.
However, processing thread might have more data to work and will loop
on each new buffer so it can delay the control operation for an
unlimited time period.
To work around this, give priority to the control operation and
do not process a new buffer in case a control operation is
pending.

Change-Id: I62fdabbb0249190d6b5639ecc47416907980cb86
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>