USB: ci13xxx_udc: Enable usb core and iface clock before reading registers
Sysfs node is present for reading the dumping registers.
But with usb cable disconnected if usb registers are dumped,
usb core and iface clocks are off, it leads to unclocked
access.

Enable and disable thse clocks before and after dumping
the registers so that unclocked access is not hit
if registers are dumped with usb cable disconnected.

Change-Id: I06f75749b8670b6331435137f65be20d5673b6c3
Signed-off-by: Sujeet Kumar <ksujeet@codeaurora.org>