msm: qpnp-power-on: change the initcall level for PON driver
Change the initcall level of PON driver from module_init() to
subsys_initcall() so that the PON devices can be available
earlier. This helps power-on/power-off reasons to be read to
debug some stability related issues. Also, it helps resources
under PON driver (e.g. regulator) to be available for other
clients.

Change-Id: Id9e0ce6b8581f23c7dc1ab2ce760969ccf10f0b6
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>