net: ipc_router: Fix use after free bug for packet type
IPC Router packet type is accessed directly with packet
structure without any locks being held. This may lead to a use
after free bug if the packet is already consumed and freed by
the client because of a previous packet arrival notification.

Access and cache the packet type within the critical section.

CRs-Fixed: 907815
Change-Id: Ice0776ea6dc3a8960b07d6c6a78de49f4ff56a9b
Signed-off-by: Atish Kumar Patra <apatra@codeaurora.org>
Signed-off-by: Andrey Cherepkov <andreiskull@gmail.com>