media: dvb: mpq: give control operations priority over processing
The processing thread might have a relatively large buffer to process
which can take a long time (a few dozen seconds at extreme cases).
During this time the processing thread holds a mutex and control
operations such as tearing down the channel have to wait until all
processing is done. To prevent this, check whether control operation
is pending during processing and release the mutex and continue to the
next iteration to allow the control operation to take place.

Change-Id: I2c98472cc466416d85a259b81ee619382dfae3f5
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>