usb: xhci: Check and free endpoint 0 ring in fail path
Previously we have backported a patch which is to prevent NULL
pointer dereference. But compared to upstream, our kernel tree
has ring_cache allocated later. This means the assumption that
eps[0].ring is not allcated incorrect.

To prevent possible leaks, we use xhci_ring_free() to properly
free it, and add it to the top of failure path.

Signed-off-by: Wang Han <416810799@qq.com>